requirements.txt
pandas~=1.1.3
pytest
pylint~=2.6.0
pydocstyle~=5.1.1
numpy~=1.19.3
matplotlib~=3.3.2
scipy~=1.5.0
opencv-python~=4.4.0.44
pillow~=8.0.1
pyinstaller
pymongo~=3.11.0
pyqt5~=5.15.1
pytesseract
virtualbox
pyyaml
requests~=2.24.0
configobj~=5.0.6
jupyter
xlib
xlrd
lmfit